# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api_rest.errorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
...
  api_rest.forgotPasswordResponse:
    properties:
      message:
        example: if the email exists, a reset link was sent
        type: string
    type: object
  api_rest.forgotRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  api_rest.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  api_rest.loginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  api_rest.logoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  api_rest.logoutResponse:
    properties:
      message:
        example: logged out
        type: string
    type: object
...
  api_rest.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  api_rest.refreshResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  api_rest.resetPasswordResponse:
    properties:
      message:
        example: password reset successful
        type: string
    type: object
  api_rest.resetRequest:
    properties:
      newPassword:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - newPassword
    - token
    type: object
...
  api_rest.signupRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  api_rest.signupResponse:
    properties:
      message:
        example: user created
        type: string
      verification_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  api_rest.verifyResponse:
    properties:
      message:
        example: email verified
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: REST, GraphQL, and WebSocket endpoints for the file service.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BalkanID File Service API
  version: "1.0"
paths:
  /api/v1/admin/audit:
    get:
      description: Retrieve system audit log with user actions and system events
      parameters:
      - description: 'Maximum number of entries to return (1-1000, default: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of entries to skip for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit log entries
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get audit log (admin only)
      tags:
      - admin
  /api/v1/admin/force-delete:
    post:
      consumes:
      - application/json
      description: Administratively force deletion of a user file or content, bypassing
        normal deletion rules
      parameters:
      - description: Force delete request with userFileId or contentId
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Force delete scheduled
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Force delete file or content (admin only)
      tags:
      - admin
  /api/v1/admin/usage:
    get:
      description: Retrieve comprehensive system usage statistics including user storage
        consumption
      produces:
      - application/json
      responses:
        "200":
          description: System usage statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get system usage statistics (admin only)
      tags:
      - admin
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.forgotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest.forgotPasswordResponse'
      summary: Forgot password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Logout payload
        in: body
        name: body
        schema:
          $ref: '#/definitions/api_rest.logoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest.logoutResponse'
      summary: Logout
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh payload
        in: body
        name: body
        schema:
          $ref: '#/definitions/api_rest.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest.refreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      summary: Refresh token
      tags:
      - auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.resetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest.resetPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      summary: Reset password
      tags:
      - auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account and returns a verification token in development
      parameters:
      - description: Signup payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.signupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api_rest.signupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      summary: Sign up
      tags:
      - auth
  /api/v1/auth/verify:
    get:
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest.verifyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      summary: Verify email
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Provide a Bearer token: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"